# libs/api-client/src/models/PracticaNueva.ts
/* tslint:disable */
/* eslint-disable */
/**
 * PreparaTAI API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { PreguntaPracticaNueva } from './PreguntaPracticaNueva';
import {
    PreguntaPracticaNuevaFromJSON,
    PreguntaPracticaNuevaFromJSONTyped,
    PreguntaPracticaNuevaToJSON,
    PreguntaPracticaNuevaToJSONTyped,
} from './PreguntaPracticaNueva';

/**
 * 
 * @export
 * @interface PracticaNueva
 */
export interface PracticaNueva {
    /**
     * 
     * @type {string}
     * @memberof PracticaNueva
     */
    estado: string;
    /**
     * 
     * @type {Array<PreguntaPracticaNueva>}
     * @memberof PracticaNueva
     */
    pregunta: Array<PreguntaPracticaNueva>;
    /**
     * 
     * @type {string}
     * @memberof PracticaNueva
     */
    nombrePractica: string;
}

/**
 * Check if a given object implements the PracticaNueva interface.
 */
export function instanceOfPracticaNueva(value: object): value is PracticaNueva {
    if (!('estado' in value) || value['estado'] === undefined) return false;
    if (!('pregunta' in value) || value['pregunta'] === undefined) return false;
    if (!('nombrePractica' in value) || value['nombrePractica'] === undefined) return false;
    return true;
}

export function PracticaNuevaFromJSON(json: any): PracticaNueva {
    return PracticaNuevaFromJSONTyped(json, false);
}

export function PracticaNuevaFromJSONTyped(json: any, ignoreDiscriminator: boolean): PracticaNueva {
    if (json == null) {
        return json;
    }
    return {
        
        'estado': json['estado'],
        'pregunta': ((json['pregunta'] as Array<any>).map(PreguntaPracticaNuevaFromJSON)),
        'nombrePractica': json['nombrePractica'],
    };
}

export function PracticaNuevaToJSON(json: any): PracticaNueva {
    return PracticaNuevaToJSONTyped(json, false);
}

export function PracticaNuevaToJSONTyped(value?: PracticaNueva | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'estado': value['estado'],
        'pregunta': ((value['pregunta'] as Array<any>).map(PreguntaPracticaNuevaToJSON)),
        'nombrePractica': value['nombrePractica'],
    };
}