# libs/api-client/src/models/PreguntaPracticaNueva.ts
/* tslint:disable */
/* eslint-disable */
/**
 * PreparaTAI API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PreguntaPracticaNueva
 */
export interface PreguntaPracticaNueva {
    /**
     * 
     * @type {string}
     * @memberof PreguntaPracticaNueva
     */
    enunciado: string;
    /**
     * 
     * @type {string}
     * @memberof PreguntaPracticaNueva
     */
    respuesta: string;
}

/**
 * Check if a given object implements the PreguntaPracticaNueva interface.
 */
export function instanceOfPreguntaPracticaNueva(value: object): value is PreguntaPracticaNueva {
    if (!('enunciado' in value) || value['enunciado'] === undefined) return false;
    if (!('respuesta' in value) || value['respuesta'] === undefined) return false;
    return true;
}

export function PreguntaPracticaNuevaFromJSON(json: any): PreguntaPracticaNueva {
    return PreguntaPracticaNuevaFromJSONTyped(json, false);
}

export function PreguntaPracticaNuevaFromJSONTyped(json: any, ignoreDiscriminator: boolean): PreguntaPracticaNueva {
    if (json == null) {
        return json;
    }
    return {
        
        'enunciado': json['enunciado'],
        'respuesta': json['respuesta'],
    };
}

export function PreguntaPracticaNuevaToJSON(json: any): PreguntaPracticaNueva {
    return PreguntaPracticaNuevaToJSONTyped(json, false);
}

export function PreguntaPracticaNuevaToJSONTyped(value?: PreguntaPracticaNueva | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'enunciado': value['enunciado'],
        'respuesta': value['respuesta'],
    };
}