# libs/contract/api/openapi.yaml
openapi: 3.1.0
info:
  title: PreparaTAI API
  version: 1.0.0
servers:
    - url: /api
      variables:
        port:
            enum:
                - '3000'
            default: '80'
paths:
  /pregunta:
    get:
      tags:
        - preguntas
      summary: Obtener preguntas
      operationId: getAllPreguntas
      responses:
        '200':
          description: Lista de preguntas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pregunta'

    post:
      tags:
        - preguntas
      summary: Guardar una pregunta
      operationId: createPregunta
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pregunta'
      responses:
        '200':
          description: Pregunta guardada con exito
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pregunta'
  /pregunta/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: el identificador de la pregunta
        schema:
          type: string
    get:
      tags:
        - preguntas
      summary: Obtener una pregunta por id
      operationId: getOnePreguntasById
      responses:
        '200':
          description: Pregunta por ID
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Pregunta'
    patch:
      tags:
        - preguntas
      summary: Actualizar una pregunta por id
      operationId: updatePreguntaById
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pregunta'
      responses:
        '200':
          description: Pregunta actualizada por ID
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Pregunta'
    delete:
      tags:
        - preguntas
      summary: Eliminar una pregunta por id
      operationId: deletePreguntaById
      responses:
        '204':
          description: Pregunta eliminada por ID
...
  /pregunta/{id}/enterrar:
    post:
      tags:
        - preguntas
      summary: Entierra la pregunta para que no salga
      operationId: enterrarPregunta
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador de la pregunta
          schema:
            type: string
      responses:
        '204':
          description: Pregunta enterrada
...
  /pregunta/porCategoria/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: el identificador de la pregunta
        schema:
          type: string
      - name: page
        in: query
        required: false
        description: Número de página (empieza en 1)
        schema:
          type: integer
          default: 1
      - name: limit
        in: query
        required: false
        description: Número de preguntas por página
        schema:
          type: integer
          default: 20
      - name: seed
        in: query
        required: false
        description: Semilla para randomizar las preguntas
        schema:
          type: integer
          default: 0
    get:
      tags:
        - preguntas
      summary: Obtener las preguntas que pertenecen a una categoria dada
      operationId: getPreguntasPorCategoria
      responses:
        '200':
          description: Pregunta que pertenecen a una categoria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pregunta'
...
  /practica/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: el identificador de la practica
        schema:
          type: string
      - name: page
        in: query
        required: false
        description: Número de página (empieza en 1)
        schema:
          type: integer
          default: 1
      - name: limit
        in: query
        required: false
        description: Número de preguntas por página
        schema:
          type: integer
          default: 20
      - name: seed
        in: query
        required: false
        description: Semilla para randomizar las preguntas
        schema:
          type: integer
          default: 0
    get:
      tags:
        - practicas
      summary: Obtener una practica por id
      operationId: getPracticaById
      responses:
        '200':
          description: Listado de preguntas de la practica
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Practica'
  /practica/crearPractica:
    post:
      tags:
        - practicas
      summary: Guardar una practica
      operationId: createPractica
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PracticaNueva'
      responses:
        '204':
          description: Practica creada
...
components:
  schemas:
    Categoria:
      type: object
      properties:
        id:
          type: string
        nombre:
          type: string
        tipo:
          type: string
      required:
        - id
        - nombre
        - tipo
...
    Respuesta:
      type: object
      properties:
        enunciado:
          type: string
        id:
          type: string
        correcta:
          type: boolean
      required:
        - enunciado
        - correcta
        - id
    Estadistica:
      type: object
      properties:
        aciertos:
          type: integer
        fallos:
          type: integer
        total:
          type: integer
      required:
        - aciertos
        - fallos
        - total
    Pregunta:
      type: object
      properties:
        id:
          type: string
        enunciado:
          type: string
        codigo:
          type: string
        estado:
          type: string
        respuestas:
          type: array
          items:
            $ref: '#/components/schemas/Respuesta'
        categorias:
          type: array
          items:
            $ref: '#/components/schemas/Categoria'
        estadisticas:
            $ref: '#/components/schemas/Estadistica'
      required:
        - id
        - enunciado
        - respuestas
        - estadisticas
        - estado
...
    PreguntaPracticaNueva:
      type: object
      properties:
        enunciado:
          type: string
        respuesta:
          type: string
      required:
        - enunciado
        - respuesta
    PracticaNueva:
      type: object
      properties:
        estado:
          type: string
        pregunta:
          type: array
          items:
            $ref: '#/components/schemas/PreguntaPracticaNueva'
        nombrePractica:
          type: string
      required:
        - pregunta
        - estado
        - nombrePractica
...
    Practica: 
      type: object
      properties:
        idPractica:
          type: string
        nombrePractica:
          type: string
        preguntas:
          type: array
          items:
            $ref: '#/components/schemas/Pregunta'
      required:
        - idPractica
        - nombrePractica
        - preguntas